+++ docker/prestart.sh
@@ -13,13 +13,11 @@ set -e
 # Ensure db is up
 echo "Waiting for db to be ready"
 python3 -m mora.cli checkdb --wait 30 || exit
-echo "OK"
 echo ""
 
 # Migrate sessiondb
 echo "Migrating sessiondb"
 python3 -m mora.cli initdb --wait 30 || exit
-echo "OK"
 echo ""
 
 # Migrate conf_db
@@ -33,11 +31,19 @@ echo ""
 # Check that DBs are up and ready
 echo "Checking db-status"
 python3 -m mora.cli check-configuration-db-status || exit
-echo "OK"
 echo ""
 
 # Wait for rabbitmq to start
 echo "Waiting for rabbitmq"
 python3 -m mora.cli wait-for-rabbitmq --seconds 30 || exit
-echo "OK"
+echo ""
+
+# Wait for object store to start
+echo "Waiting for minio"
+python3 -m mora.cli wait-for-minio --seconds 30 || exit
+echo ""
+
+# Ensure object store bucket exists
+echo "Ensuring minio bucket exists"
+python3 -m mora.cli ensure-minio-bucket --seconds 30 || exit
 echo ""